Small React Project
This is the core functionality of the React project. The project creates a form that
a user can fill out. Once filled out, the user's entries are relayed back to them through
an alert command. The entries are seemingly stored in two sets of useState() Hooks, but that
is somewhat misleading. The first set of hooks is utilized to set the form field values in the
alert call. After the hooks have been parsed into the call, they are reinitialized so the user
can fill the form out again. The second set of hooks persists the form field values and does
not get reinitialized. This is how the recall button operates.